fix: feat: define IErrorPage interface and update ErrorPage component
