+++ src/types/design/IErrorPage/index.ts
@@ -1,0 +1,8 @@
+interface IErrorPage {
+  errorCode?: number;
+  heading?: string;
+  nameButton?: string;
+  onClick?: () => void;
+}
+
+export type { IErrorPage };
